Simplify setup of model in one class
Now we have functions for the different interaction terms
as well as a class to manage all resources together for
the 2D HP lattice simulations. One can initialize at one
place, e.g.,
model = Lattice_HP_QUBO(dim = [10, 10], sequence="HPHPHPHPHPH")
All the initialisation steps will be carried out in the proper
order. The dictionary containing Q will be generated, as will
the key list, so that the bit vectors can be properly
interpreted. Some basic configuration mistakes will be caught,
such as the box being too small for the sequence. The lambda
parameters can be specified as another argument to the constructor.
But, if not specified, they will be set to the known best values
for different sequence sizes.